cloudinary configuration

+++ src/api/v1/controllers/fruits.controller.js
@@ -1,5 +1,10 @@
 const Fruit = require("../models/fruits.model")
 
+const Cloudinary = require("cloudinary").v2;
+
+const DatauriParser = require("datauri/parser");
+const parser = new DatauriParser();
+
 const getFruits = async (req,res,next) => {
     try{
         const fruits = await Fruit.find();
@@ -27,7 +32,30 @@ const getFruitById = async (req,res,next) => {
 
 const createFruit = async (req,res,next) => {
     try{
-
+        const {name,price} = req.body;
+        const inputImage = req.file;
+        let imageUrl = '';
+        if(inputImage != null){
+            const imageFile = inputImage.buffer;
+            const result = parser.format(".png", imageFile);
+            const imageData = await Cloudinary.uploader.upload(result.content, {
+                folder :"Fruits",
+                use_filename : "true",
+                unique_filename : "false"
+            });
+            imageUrl = imageData.secure_url; 
+            console.log(imageUrl)
+        }
+        const newFruit = await Fruit.create({
+            name: name,
+            price: price,
+            imageUrl: imageUrl
+        });
+        console.log(newFruit)
+        return res.status(201).json({
+            message: "Fruit created successfully",
+            data: newFruit
+        })
     }catch(error){
         next(error)
     }
@@ -35,6 +63,25 @@ const createFruit = async (req,res,next) => {
 
 const updateFruit = async (req,res,next) => {
     try{
+        const {name,price} = req.body;
+        let updates = {};
+        const fruitId = req.params.id;
+        const fruit = await Fruit.findById(fruitId);
+        if(fruit == null){
+            return res.status(404).json({
+                message: "Fruit not found"
+            })
+        }
+        if(req.file){
+            const imageFile = req.file.buffer;
+            const result = parser.format(".png", imageFile);
+            const imageData = await Cloudinary.uploader.upload(result.content, {
+                folder :"Fruits",
+                use_filename : true,
+                unique_filename : false
+            });
+            updates.imageUrl = imageData.secure_url; 
+        }
 
     }catch(error){
         next(error)
@@ -44,6 +91,12 @@ const updateFruit = async (req,res,next) => {
 const deleteFruit = async (req,res,next) => {
     try{
         const fruitId = req.params.id;
+        const fruit = await Fruit.findById(fruitId);
+        if(fruit == null){
+            return res.status(404).json({
+                message: "Fruit not found"
+            })
+        }
         const deletedFruit = await Fruit.findByIdAndDelete(fruitId);
         return res.status(200).json({
             message: "Fruit deleted successfully",
